:recycle: Moved progressBar1 into separate file
Moved progressBars into their own folder to add better structure
